DatabaseWrapper.Core.DatabaseHelperBase: New public method: ExpresionToWhereClause. Thanks to using sql parameters it is almost generic now.
